Major project-wide dev update, cleanup & refact
added long-forgotten/missing .h/.cpp files to repo;
updated PBASLBSP to only use rel LBSP features and to toggle back to abs (if needed);
rewrote PBASLBSP gradient prop func for color sum distances;
replaced util LBSP rel/abs func by general impl (must pass specific thrsh each time);
added 1ch1t/1ch3t variations of low-level LBSP calc files;
cleaned up PBASLBSP/VIBELBSP algorithms so both use common values via const variables;
added mouse-driven debug output to PBASLBSP debug frames;
added rel/abs LBSP cutoff feature in PBASLBSP (to be tested, still sub-optimal);
replaced all int-based LBSP thresholds by uchar's;
added support to only process sequence-ROI-approved keypoints for PBASLBSP/ViBeLBSP;
added FPS calc/output to log/results files;

+++ BackgroundSubtractorViBeLBSP.h
@@ -30,7 +30,7 @@ public:
 	//! default constructor (also uses the default LBSP descriptor extractor constructor & params)
 	BackgroundSubtractorViBeLBSP();
 	//! full constructor used to intialize an 'absolute' LBSP-based background subtractor
-	BackgroundSubtractorViBeLBSP(	int nLBSPThreshold,
+	BackgroundSubtractorViBeLBSP(	uchar nLBSPThreshold,
 									int nDescDistThreshold=BGSLBSP_DEFAULT_DESC_DIST_THRESHOLD,
 									int nColorDistThreshold=BGSVIBELBSP_DEFAULT_COLOR_DIST_THRESHOLD,
 									int nBGSamples=BGSVIBELBSP_DEFAULT_NB_BG_SAMPLES,
